[bug][feature] Load-legacy into secondary root
In case of failure (most likely because the preset has the new JSON
format) the preset would otherwise be empty, if the JSON load also
fails. In case no load variant is successful the expectation is that
the previous preset is kept (and possibly an error message shown).

Now only load if prerequisites are met, i.e. file magic & version.
If so, load into secondary and then swap locations of secondary and
primary root.

This has the added side effect that loading can take place while the
old preset is still rendering. Only the swap now requires the render
lock. Loading shouldn't take that long, but it might still interfere
slightly. The swap should be much faster than any load operation.