[api][audio] Set audio end time in samples
When matching manual audio input to the existing data, we need a
reference timestamp. We use the sample index of the last entry in the
given data. This is an easy-to-calculate-with integer value. And if
the sample rate changes, we simply throw away the whole pre-existing
buffer. The samplerate should be pretty stable over time and only
change occasionally.