[api] Change avs_render_frame()'s time_ms meaning
Previously the value of time_ms split at +1000 between realtime and
video mode. This was confusing and a bit arbitrary.

Change the type to signed 64-bit integer, and the interpretation to
mean <0 for realtime and >=0 for video mode. The loss of one bit only
halves the maximum expressible time range from from 580 to 290 million
years -- both ludicrous.

Add a few paragraphs to the doc for avs_render_frame() explaining what
the assumptions behind realtime- and video mode are.

Change the name of the parameter from time_ms to time_in_ms.