[effect] All effects gain a comment property
Any effect type may warrant some additional information. This change
enables a unified way for preset authors to provide them. The comment
is now the same kind of built-in effect parameter like "enabled".

For the Comment effect it replaces its only config parameter and
Comment is now a "parameterless" effect (i.e. it has only the default
properties, "enabled" and "comment").

Since this moves out of the "config" section of the effect, this does
pose the question of how to expose this through the API?
    Option 1: Expose as additional get/set functions like for
              "enabled".
    Option 2: Create a new "default properties" struct that one can
              retrieve for any component, containing "enabled" &
              "comment".
Option 2 does have the benefit of being easily extendable in the
future, should other default properties be introduced. It would extend
the API by a struct instead of 2 functions, which would be the cost of
Option 1.