[build] Split common/win/linux files and flags
ifdef-guard the bodies of getmouse() and gettime(), which are not yet
available in libavs since it doesn't know about the host application.
I haven't thought about the API design for that yet.

Also, compile an intermediate "avs_common" object for reuse by other
build targets, currently libavs and vis_avs.