[ci] Skip clang-format for disabled directories
Search for ".clang-format" files with "DisableFormat" lines and remove
any files whose path is below their location (i.e. their path with the
filename removed).

This reduces the runtime of the clang-format CI step by ~5 minutes,
because it takes that long for clang-format to work through the json.h
library header.