[port] Account for platform-specific lib filenames
On Linux library names have a "lib" prefix and the ".so" extension
must be included in the name.

Windows libraries (DLLs) usually don't have a "lib" prefix. Also when
loading the lib by name the ".dll" extension is not included.
The ffmpeg libs include their major version number in the name too.