[refactor] Move `line()` declaration to own header
Currently we want to replace r_defs.h and the blend functions in most
or all places. The line drawing function is also ripe for refactoring
(especially with anti-aliasing) but not right now.

Extract its declaration so that r_defs.h includes can be removed and,
if line drawing is needed, replaced with a linedraw.h include. Some
type coercions are necessary to fit the new blendmode functions'
uint32_t input type for the framebuffer.