[refactor] Blend functions: Split src2 & dest
The initial blend function interface was too optimistic:

While most cases blend some buffer with its destination buffer,
sometimes the source buffer is blended with a second source into the
destination buffer (i.e. what was previously in the destination is
not directly considered at all).

So expand the blend function interface from
    *dest = *src @ *dest
to
    *dest = *src1 @ *src2
where @ is the blend operation and in most cases src2 is the same
address as dest.