[doc] Increase minimum supported GCC version to 9
Since the legacy GCC docker image doesn't (readily) support the
required CMake version 3.18, the CI was updated to use GCC 9.

GCC 8 may continue to work, and one could possibly add CMake 3.18 to
the old Debian image to make a CI for it work again, but it's just not
worth the effort at this time: GCC 9 was released in 2019, so it's old
enough to be widely supported. Just as, or even more so than, GCC 8
was when this project started.