[dbg] Better video/libav load error message
Some libav libs (not all) happened to fail to load. The reason was
that a dependent library had been upgraded but the ffmpeg/libav
package had not yet been rebuilt. This was not visible in an error
message, instead the video effect silently failed to render anything.

Several changes address this:
- Log a warning on video file load in the Video effect.
- Add the system error from the dlib loader to the error from the
  libav loader.
- Upgrade the libav helper class error message to a std::string for
  more readable string concat code.
- Add a `library_error()` method to the platform dlib util methods.