[bug] Fix preset (re)load secondary root swap
A preset is loaded into the secondary Root of an instance first, so
that rendering of the old preset can continue during loading. Then the
two Roots are swapped out, and the formerly secondary Root is used for
rendering.

Using `std::swap` could not work because the Effect was not publicly
swappable. Make Effect's `swap()` method public and use that method
directly.

Also, clean the secondary right before loading. Before, just straight
loading the new preset appended the new to the previous preset in the
secondary. This led to a duplication of effects starting after the
preset was (re)loaded twice.