[api][feature] Add JSON schema generation
The JSON schema is only generated as information for others and no
preset files are validated with it by AVS. Other tools (such as UIs)
however may preemptively warn & point at potential problems in a file.

AVS could theoretically make use of the schema itself. But the
dependency complexity of including a JSON schema validation library
would not be worth the feature. Moreover it's arguably more desirable
from a user standpoint to be able to load imperfect presets even
partially, than reject them on formality.