[bug] Get missing video length for some containers
The frame count & duration of a video are only header metadata, and
optional. Notably, the Matroshka/MKV container format records neither
frame count nor duration of the video stream. An AVS_Video instance
from such a file had zero length and was unplayable.

To fix this it's actually surprisingly quick to let the demuxer run
through all the frames and add up the frame count and the frames'
duration. A two-hour test video was scanned in <1s on a 4.5GHz laptop
CPU.

A 1s load time is noticeable but acceptable and actual videos will
rarely be this long. So no explicit async waiting/in-decoding-state
logic was implemented and the code left inline during the construction
of the AVS_Video object. But an info-level log line is emitted if this
manual frame/duration counting is done.