[bug] Videos are actually mostly _not_ upside-down
The assumption before was that (for some unknown reason) the libav
decoders would return the frame upside down, always. As it turns out,
this is specific to the video codec & file and the decoder will
provide this information for every frame.

`AVS_Video`'s `copy()` and `copy_with_aspect_ratio()` methods will now
only vertically invert the frame when it is actually upside-down as
reported by the decoder.

Curiously, the "Video" effect corrected the wrongly inverted frame,
and inverted it _again_. Fix that too.