[feature] Add Rust lib wrapper & standalone client
This is my current Linux-native testbed.

Split the single-file version binding+CLI file into proper "avs-rs"
and "avs-cli" modules. Use bindgen to generate the basic Rust bindings
from the headers and create a more idiomatic interface on top of it.

Reload the preset automatically when the preset file changes. In the
absence of a proper editor UI & with a text editor this is as good as
it gets in terms of editing presets outside of Winamp for now.

In the presence of a 32-bit Rust toolchain and if 32-bit packages are
installed under /usr/lib32/, running the build via CMake should build
the Rust code on Linux along with a normal build.