[audio][linux] Make libpipewire optional
Internal audio input (where AVS opens its own audio input device) is
only one of two audio modes in AVS. Audio data can also be provided
"externally" with the `avs_set_audio()` API. So depending on Pipewire
by linking it in on Linux is needlessly restrictive (especially for a
32bit library, when few distros readily provide a 32bit libpipewire).

Instead, load it at runtime now, like the ffmpeg/libav libs for video
support.

The `spa_*()` utility functions from Pipewire's own SPA plugin lib are
not part of the pipewire ABI and need not be in the lib wrapper, since
the SPA utils are a header-only library.