[doc] Add "bytes" comment to srli_si128 asm calls
Most shift operations take value in bits. But `_mm_srli_si128()`,
a.k.a. `PSRLDQ`, takes a byte value instead, which can make the code
confusing to read, if you don't know about it.

Add comments with the unit to the shift operation, as well as one
"bits" unit to a preceding "normal" shift, to make things clear.