[bug] Fix darken modes in SIMD code of Multiplier
`_mm_srli_si128()`, a.k.a. `PSRLDQ` shifts by bytes instead of bits.
So we need to use another one, any of the 16bit-, 32bit- or 64bit
packed integer shifts would work, but we choose the 32bit simply
because RGB0_8 pixels are that size.

What we would really need is an 8bit packed integer shift, then we
could remove the masking AND afterwards, but 8bit packed integer
support is sadly very incomplete in the x86 SIMD instruction sets.

We leave the masks below as-is, even though we don't technically need
to mask all of the bits unset in them, but it's just more readable
that way.