[bug][eel] Use AVSTrans' linear mode for EEL2
AVSTrans other modes cannot deal with EEL2's compound assignment
operators ("+=", "-=", etc.) and instead parse them as equals with a
rogue symbol next to the lhs. This transformed
    a += 1;
into
    assign(a +, 1);

The "linear" mode in AVSTrans does not turn statements into
expressions and thus leaves these compound assignments as-is.

(Also clang-formats a line earlier in the file.)