[feature] Support multiple instances with AVSTrans
Move the storage location for the patched-in EEL pre-/post-compile
hooks from global to `compileContext` (`NSEEL_VMCTX` externally).

AVSTrans now makes use of instance-aware global config and sets its
hooks for the instance, which in turn forwards setting the hooks to
EEL on every code compile of every programmable effect in the AVS
instance.

Remove the `NSEEL_unset_compile_hooks()` function: Just use the setter
with NULLs to unset. This way the hooks are automatically unset on
every subsequent compile call if AVSTrans is removed from the preset.

Make `Configurable_Effect::globals` private as it should be. Effects
should only be able to access their `global` instance, not those of
other AVS instances (at least this is not something currently wanted).

Update the EEL hooks changes in the dependency patch file accordingly.