[clean] Rename Root "author" config to "contributor"
This is a bit of a personal (almost political) commit. I have no great
fondness for the "Author" concept, and "Contributor" seems like a more
general, open term.

AVS always allowed for a lot of non-goal-oriented experimentation with
its effect's parameters. It's one of the endearing features of the
classic UI, in my opinion. In that sense I often felt that many were
"curators" of presets, i.e. "selecting works from a pool of
possibilities", rather than "authors". However, there sure are plenty
of presets that were created very purposefully and with a lot of
forethought.