[ci] Add CI tasks script
The commands and flags used to run AVS have become quite involved over
time. This is a) a barrier to entry for newcomers, and b) convolutes
CI configs by mixing build details with each CI's own peculiarities.
To separate this, document those commands in this file.

Use these in CI configs in the future.