[build] Use 64bit package headers for runtime libs
AVS tries to load heavy libraries at runtime only so that AVS may be
built and run without them, although with a reduced feature-set of
course. Currently this includes FFmpeg and (on Linux only) Pipewire.

Since AVS is still 32bit-only software, and 32bit support is waning,
this can make building on modern Linux systems difficult, because the
required libraries are only available in 64bit versions. At build time
only the headers are needed though, and these are (for the current set
of libs at least) the same for both platforms. So AVS can use the
headers available from the 64bit packages.

To do this, create a CMake macro that checks for the 32bit libs via
pkg-config first and, if it can't be found, modifies the search path
for pkg-config to common 64bit paths and searches again. Only do this
at all on Linux.