[feature] Add warmup frames to Root config
It's useful to render screenshots of presets for views of preset
collections in folders and websites. When generating a screenshot of a
preset it's often important to "warm up" the preset a bit (i.e. render
a few frames) before taking the picture, to get a representative and
interesting view. What the best number of warmup frames is, depends on
the preset, obviously. Some need none, others a lot. The safe thing to
do is to always render a lot of warmup frames which is often unneeded.

To facilitate AVS hosts which generate such previews presets can now
optionally declare the desired number of warmup frames. AVS does not
make use of this internally. Hosts should retrieve the value from the
Root component and act accordingly.

The default for new presets is 10, a low default number so that the
average thumbnail creation time is low. Legacy presets get their value
set to -1, i.e. "unknown" and it's up to the host to choose the number
of frames.