[refactor] Store Parameter handle outside of itself
Parameters used to store their own handle as a member. The handle was
calculated at compile time using a hash of the name. Including more
information, such as the effect name and (in case of nested effects)
the parameter-tree path was not feasible, or perhaps possible. Hashing
only the name has one minor common and one larger but rare drawback:

  - Parameter names must be unique within effects (this is achievable)
  - Nested parameter info collections cannot be reused, because then
    the name occurs multiple times at each use (this is annoying)

The latter was the reason for the refactor where:

  - Parameter handles are now stored in a global map, just as effect
    handles are.
  - Parameter handles are now assigned at runtime when AVS is
    initialized and the effect library generated.
  - Parameter handles are now hashed from the effect name + the full
    path of the parameter.

It loses the nice compile-time handle setting, but makes some of the
code less cumbersome and C++-y.