[win][ui][bug] Fix new component insert position
When a normal (no children) component is selected, inserting a new
component failed, because INSERT_CHILD is not allowed in that case.

Fix by only inserting as child if the selected component can actually
have child components. Also change UI tree insertion position to
match: Before, inserting an element into a child-capable item would
add it to the beginnning of the child list. Now the UI follows the
libavs behavior of inserting at the end of the list.